# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "房屋租赁管理系统 API",
    "version": "1.0.0",
    "description": "房屋租赁管理系统前后端交互契约"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/complaint": {
      "post": {
        "tags": ["complaint"],
        "summary": "提交投诉",
        "operationId": "create_complaint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/ComplaintRecord" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/complaint/list": {
      "get": {
        "tags": ["complaint"],
        "summary": "获取投诉列表",
        "operationId": "get_complaint_list",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/ComplaintStatus" } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "allOf": [
                            { "$ref": "#/components/schemas/PaginatedResponse" },
                            {
                              "type": "object",
                              "properties": {
                                "items": {
                                  "type": "array",
                                  "items": { "$ref": "#/components/schemas/ComplaintRecord" }
                                }
                              },
                              "required": ["items"]
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/complaint/handle": {
      "post": {
        "tags": ["complaint"],
        "summary": "处理投诉（管理员）",
        "operationId": "handle_complaint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandleComplaintRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/ComplaintRecord" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "业务状态码：200成功，4xx客户端错误，5xx服务端错误"
          },
          "message": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "description": "业务数据",
            "nullable": true,
            "oneOf": [
              { "type": "object", "additionalProperties": true },
              { "type": "array", "items": {} }
            ]
          }
        },
        "required": ["code", "message"]
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "总条数"
          },
          "page": {
            "type": "integer",
            "description": "当前页码"
          },
          "page_size": {
            "type": "integer",
            "description": "每页条数"
          },
          "items": {
            "type": "array",
            "description": "当前页数据",
            "items": {}
          }
        },
        "required": ["total", "page", "page_size", "items"]
      },
...
      "ComplaintType": {
        "type": "string",
        "enum": ["house", "landlord", "other"],
        "description": "投诉类型"
      },
      "ComplaintStatus": {
        "type": "string",
        "enum": ["pending", "resolved"],
        "description": "投诉状态"
      },
...
      "ComplaintRequest": {
        "type": "object",
        "properties": {
          "type": { "$ref": "#/components/schemas/ComplaintType" },
          "content": { "type": "string", "minLength": 1, "maxLength": 2000 },
          "landlord_id": { "type": "integer", "description": "被投诉的房东ID（type=landlord 时必填）" },
          "house_id": { "type": "integer", "description": "被投诉的房源ID（type=house 时必填，房东由房源反推）" }
        },
        "required": ["type", "content"]
      },
      "ComplaintRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "tenant_id": { "type": "integer" },
          "tenant_nickname": { "type": "string" },
          "type": { "$ref": "#/components/schemas/ComplaintType" },
          "type_label": { "type": "string" },
          "content": { "type": "string" },
          "status": { "$ref": "#/components/schemas/ComplaintStatus" },
          "status_label": { "type": "string" },
          "admin_feedback": { "type": "string", "default": "" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "HandleComplaintRequest": {
        "type": "object",
        "properties": {
          "complaint_id": { "type": "integer" },
          "feedback": { "type": "string", "minLength": 1, "maxLength": 2000 }
        },
        "required": ["complaint_id", "feedback"]
      },
...
    }
  },
...
}